Testing/CSV_Datei_Automation/requirements.txt
requests==2.31.0
pandas==2.0.3
watchdog==3.0.0
python-decouple==3.8
schedule==1.2.0
